+++ src/tests/transaction-flow.ts
@@ -150,7 +150,7 @@ console.log('');
 console.log('Testing network auxiliary functions do not throw');
 await testLocalAndRemote(async () => {
   await assert.doesNotReject(async () => {
-    await Mina.transaction({ sender, fee: transactionFee }, () => {
+    await Mina.transaction({ sender, fee: transactionFee }, async () => {
       Mina.getNetworkConstants();
       Mina.getNetworkState();
       Mina.getNetworkId();
@@ -181,7 +181,7 @@ await testLocalAndRemote(async () => {
   await assert.doesNotReject(async () => {
     const transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         zkApp.deploy({ verificationKey });
       }
     );
@@ -198,7 +198,7 @@ await testLocalAndRemote(async () => {
   await assert.doesNotReject(async () => {
     const transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         zkApp.update(Field(1), PrivateKey.random().toPublicKey());
       }
     );
@@ -220,7 +220,7 @@ await testLocalAndRemote(async () => {
   await assert.doesNotReject(async () => {
     const transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         zkApp.update(Field(1), PrivateKey.random().toPublicKey());
       }
     );
@@ -238,7 +238,7 @@ console.log(
 await testLocalAndRemote(async () => {
   const transaction = await Mina.transaction(
     { sender, fee: transactionFee },
-    () => {
+    async () => {
       AccountUpdate.fundNewAccount(zkAppAddress);
       zkApp.update(Field(1), PrivateKey.random().toPublicKey());
     }
@@ -256,7 +256,7 @@ await testLocalAndRemote(async () => {
   await assert.rejects(async () => {
     const transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         AccountUpdate.fundNewAccount(zkAppAddress);
         zkApp.update(Field(1), PrivateKey.random().toPublicKey());
       }
@@ -272,7 +272,7 @@ await testLocalAndRemote(async () => {
   try {
     let transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         zkApp.incrementCounter();
       }
     );
@@ -281,7 +281,7 @@ await testLocalAndRemote(async () => {
 
     transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         zkApp.rollupIncrements();
       }
     );
@@ -290,7 +290,7 @@ await testLocalAndRemote(async () => {
 
     transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         zkApp.incrementCounter();
         zkApp.incrementCounter();
         zkApp.incrementCounter();
@@ -303,7 +303,7 @@ await testLocalAndRemote(async () => {
 
     transaction = await Mina.transaction(
       { sender, fee: transactionFee },
-      () => {
+      async () => {
         zkApp.rollupIncrements();
       }
     );